RUBY-1033: Move builders into own namespace
