+++ spec/mongo/collection/view/readable_spec.rb
@@ -770,81 +770,5 @@ describe Mongo::Collection::View::Readable do
         expect(view.sort).to eq(options[:sort])
       end
     end
-
-    context 'when an option is a cursor flag' do
-
-      let(:query_spec_options) do
-        view.send(:query_spec)[:options]
-      end
-
-      context 'when allow_partial_results is set as an option' do
-
-        let(:options) do
-          { :allow_partial_results => true }
-        end
-
-        it 'sets the cursor flag' do
-          expect(query_spec_options[:flags]).to eq([:partial])
-        end
-
-        context 'when allow_partial_results is also called as a method' do
-
-          before do
-            view.allow_partial_results
-          end
-
-          it 'sets only one cursor flag' do
-            expect(query_spec_options[:flags]).to eq([:partial])
-          end
-        end
-      end
-
-      context 'when oplog_replay is set as an option' do
-
-        let(:options) do
-          { :oplog_replay => true }
-        end
-
-        it 'sets the cursor flag' do
-          expect(query_spec_options[:flags]).to eq([:oplog_replay])
-        end
-      end
-
-      context 'when no_cursor_timeout is set as an option' do
-
-        let(:options) do
-          { :no_cursor_timeout => true }
-        end
-
-        it 'sets the cursor flag' do
-          expect(query_spec_options[:flags]).to eq([:no_cursor_timeout])
-        end
-      end
-
-      context 'when cursor_type is set as an option' do
-
-        context 'when :tailable is the cursor type' do
-
-          let(:options) do
-            { :cursor_type => :tailable }
-          end
-
-          it 'sets the cursor flag' do
-            expect(query_spec_options[:flags]).to eq([:tailable_cursor])
-          end
-        end
-
-        context 'when :tailable_await is the cursor type' do
-
-          let(:options) do
-            { :cursor_type => :tailable_await }
-          end
-
-          it 'sets the cursor flags' do
-            expect(query_spec_options[:flags]).to eq([:await_data, :tailable_cursor])
-          end
-        end
-      end
-    end
   end
 end
